fix(mcp): keep spawn validation handler out of route exports
Move the validation implementation behind the route handler so Next production type generation accepts the new endpoint.
